Provide helpful errors on JSON.parse exceptions
The current code was using `jsonlint` as the default parser. I replaced
it with `json-parse-helpfulerror` which uses JSON.parse by default and
only if it throws it uses a different parser to provide helpful errors
so in most cases the performance will be better. The alternative parser
it uses is called `jju` and it provides really good clues to where the
issue is.

Other than that, I also added a test script to package.json, so you can
run `npm test` to run the tests and updated the license field to use the
valid format (without this, when ever you install cjson, npm will show a
warning message that the package is missing the license field).